refactor: 도메인별 패키지를 계층형 구조로 변경
