+++ src/main/java/roomescape/domain/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.reservationTime;
+package roomescape.domain;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
