+++ src/main/java/roomescape/repository/TimeUpdatingRepository.java
@@ -1,17 +1,18 @@
-package roomescape.reservationTime;
+package roomescape.repository;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
 
 import java.sql.PreparedStatement;
 
 @Repository
-public class TimeUpdatingDao {
+public class TimeUpdatingRepository {
     private JdbcTemplate jdbcTemplate;
 
-    public TimeUpdatingDao(JdbcTemplate jdbcTemplate) {
+    public TimeUpdatingRepository(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
